requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
passporteye
Pillow
pytesseract
opencv-python-headless
requests
selenium
webdriver_manager
httpx
python-dotenv
pydantic[email]